requirements.txt
streamlit
pandas
matplotlib
openpyxl
reportlab
Pillow
